+++ src/Ping/Fallback.php
@@ -3,10 +3,9 @@ declare(strict_types = 1);
 
 namespace Innmind\FileWatch\Ping;
 
-use Innmind\FileWatch\Ping;
 use Innmind\Immutable\Attempt;
 
-final class Fallback implements Ping
+final class Fallback implements Implementation
 {
     private ProcessOutput $attempt;
     private OutputDiff $fallback;