+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Innmind\FileWatch\Ping::__invoke()` now returns an `Innmind\Immutable\Attempt`
 - `Innmind\FileWatch\Watch` is now a final class
+- `Innmind\FileWatch\Ping` is now a final class
 
 ### Fixed